+++ src/Provider/ApiPlatform/IriProvider.php
@@ -4,8 +4,6 @@ declare(strict_types=1);
 
 namespace AutoMapper\Provider\ApiPlatform;
 
-use ApiPlatform\Api\IriConverterInterface as LegacyIriConverterInterface;
-use ApiPlatform\Api\ResourceClassResolverInterface as LegacyResourceClassResolverInterface;
 use ApiPlatform\Metadata\IriConverterInterface;
 use ApiPlatform\Metadata\ResourceClassResolverInterface;
 use AutoMapper\MapperContext;
@@ -15,8 +13,8 @@ use AutoMapper\Provider\ProviderInterface;
 final readonly class IriProvider implements ProviderInterface
 {
     public function __construct(
-        private LegacyIriConverterInterface|IriConverterInterface $iriConverter,
-        private LegacyResourceClassResolverInterface|ResourceClassResolverInterface $resourceClassResolver,
+        private IriConverterInterface $iriConverter,
+        private ResourceClassResolverInterface $resourceClassResolver,
     ) {
     }