+++ src/Transformer/ArrayToDoctrineCollectionTransformer.php
@@ -26,7 +26,7 @@ final class ArrayToDoctrineCollectionTransformer implements \Stringable, Transfo
     ) {
     }
 
-    public function transform(Expr $input, Expr $target, PropertyMetadata $propertyMapping, UniqueVariableScope $uniqueVariableScope, Expr\Variable $source, ?Expr $existingValue = null): array
+    public function transform(Expr $input, Expr $target, PropertyMetadata $propertyMapping, UniqueVariableScope $uniqueVariableScope, Expr $source, ?Expr $existingValue = null): array
     {
         /**
          * $collection = new ArrayCollection();.