+++ src/Transformer/SourceEnumTransformer.php
@@ -17,7 +17,7 @@ use PhpParser\Node\Expr;
  */
 final class SourceEnumTransformer implements TransformerInterface
 {
-    public function transform(Expr $input, Expr $target, PropertyMetadata $propertyMapping, UniqueVariableScope $uniqueVariableScope, Expr\Variable $source, ?Expr $existingValue = null): array
+    public function transform(Expr $input, Expr $target, PropertyMetadata $propertyMapping, UniqueVariableScope $uniqueVariableScope, Expr $source, ?Expr $existingValue = null): array
     {
         /* $input->value */
         return [new Expr\PropertyFetch($input, 'value'), []];