+++ tests/Bundle/Resources/App/Api/Entity/ShelfMapped.php
@@ -6,6 +6,7 @@ namespace AutoMapper\Tests\Bundle\Resources\App\Api\Entity;
 
 use ApiPlatform\Metadata\ApiResource;
 use ApiPlatform\Metadata\Get;
+use AutoMapper\Attribute\Mapper;
 use AutoMapper\Tests\Bundle\Resources\App\Api\Provider\ShelfProvider;
 use Doctrine\Common\Collections\Collection;
 use Symfony\Component\Serializer\Attribute\Groups;
@@ -13,18 +14,19 @@ use Symfony\Component\Serializer\Attribute\Groups;
 #[ApiResource(
     operations: [
         new Get(
-            uriTemplate: '/shelf',
+            uriTemplate: '/shelf-mapped',
             normalizationContext: ['groups' => ['shelf']],
             provider: ShelfProvider::class,
         ),
         new Get(
-            uriTemplate: '/shelf-group',
+            uriTemplate: '/shelf-mapped-group',
             normalizationContext: ['groups' => ['group']],
             provider: ShelfProvider::class,
         ),
     ],
 )]
-class Shelf
+#[Mapper(source: 'array', target: 'array')]
+class ShelfMapped
 {
     public function __construct(
         #[Groups(['shelf', 'group'])]